+++ src/stelnet/board/contact/DisplayContact.java
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 import stelnet.CommonL10n;
 import stelnet.util.L10n;
 import stelnet.util.StringUtils;
-import stelnet.widget.HeadingWithButtons;
+import stelnet.widget.heading.HeadingWithButtons;
 import uilib.UiConstants;
 import uilib.property.Size;
