+++ src/stelnet/board/query/grouping/SystemView.java
@@ -9,9 +9,9 @@ import stelnet.board.query.QueryL10n;
 import stelnet.board.query.ResultIntel;
 import stelnet.board.query.ResultSet;
 import stelnet.board.query.ResultView;
-import stelnet.board.query.view.result.MarketHeader;
 import stelnet.util.CargoUtils;
 import stelnet.util.L10n;
+import stelnet.widget.heading.MarketHeader;
 import uilib.Renderable;
 import uilib.ShowCargo;
 import uilib.ShowShips;
